+++ CHANGELOG.md
@@ -9,8 +9,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Support tilde expansion (`~`) to home directory for definition and documentation file paths
+- Added a datamodel-aware `WaitForChild` function
 - We now apply a datamodel-aware `FindFirstChild` function to the top level datamodel and service types ([#543](https://github.com/JohnnyMorganz/luau-lsp/issues/543))
-- Added autocompletion of children to `:FindFirstChild("")` ([#685](https://github.com/JohnnyMorganz/luau-lsp/issues/685))
+- Added autocompletion of children to `:FindFirstChild("")` and `:WaitForChild("")` ([#685](https://github.com/JohnnyMorganz/luau-lsp/issues/685))
 
 ### Changed